+++ packages/fiori-freestyle-writer/test/basic.test.ts
@@ -6,6 +6,10 @@ import { testOutputDir, debug, updatePackageJSONDependencyToUseLocalPath } from 
 import { OdataVersion, ServiceType } from '@sap-ux/odata-service-writer';
 import type { BasicAppSettings } from '../src/types';
 import { projectChecks } from './common';
+import { applyCAPUpdates, type CapServiceCdsInfo } from '@sap-ux/cap-config-writer';
+import { create as createStorage } from 'mem-fs';
+import { create, type Editor } from 'mem-fs-editor';
+import { types } from 'util';
 
 const TEST_NAME = 'basicTemplate';
 jest.setTimeout(240000); // Needed when debug.enabled
@@ -17,6 +21,11 @@ jest.mock('read-pkg-up', () => ({
             version: '9.9.9-mocked'
         }
     })
+}));
+
+jest.mock('@sap-ux/cap-config-writer', () => ({
+    ...jest.requireActual('@sap-ux/cap-config-writer'),
+    applyCAPUpdates: jest.fn()
 }));
 
 describe(`Fiori freestyle template: ${TEST_NAME}`, () => {
@@ -362,6 +371,110 @@ describe(`Fiori freestyle template: ${TEST_NAME}`, () => {
                     manifest.json['sap.ui5'].routing.targets[`Target${viewPrefix}`].viewName
                 ].every((entry) => entry.includes(viewPrefix))
             ).toBeTruthy();
+        });
+    });
+
+    describe('CAP updates', () => {
+        const capService: CapServiceCdsInfo = {
+            cdsUi5PluginInfo: {
+                isCdsUi5PluginEnabled: true,
+                hasMinCdsVersion: true,
+                isWorkspaceEnabled: true,
+                hasCdsUi5Plugin: true
+            },
+            projectPath: 'test/path',
+            serviceName: 'test-service',
+            capType: 'Node.js'
+        };
+
+        const getFreestyleApp = (options: { typescript: boolean; sapux: boolean; capService?: CapServiceCdsInfo }) => {
+            const { typescript, sapux, capService } = options;
+            return {
+                app: {
+                    id: 'ff-basic-id'
+                },
+                template: {
+                    type: TemplateType.Basic,
+                    settings: {}
+                },
+                service: {
+                    version: OdataVersion.v4,
+                    capService
+                },
+                package: {
+                    name: 'ff-basic-id'
+                },
+                appOptions: {
+                    sapux,
+                    typescript
+                }
+            } as FreestyleApp<BasicAppSettings>;
+        };
+
+        const capProjectSettings = {
+            appRoot: curTestOutPath,
+            packageName: 'ff-basic-id',
+            appId: 'ff-basic-id',
+            sapux: true,
+            enableTypescript: false
+        };
+
+        afterEach(() => {
+            jest.clearAllMocks();
+            jest.resetAllMocks();
+        });
+
+        test('should perform CAP updates when CAP service is available and cds ui5 plugin is enabled', async () => {
+            const fs = create(createStorage());
+
+            const freestyleApp: FreestyleApp<BasicAppSettings> = getFreestyleApp({
+                sapux: false,
+                typescript: false,
+                capService
+            });
+            await generate(curTestOutPath, freestyleApp, fs);
+            expect(applyCAPUpdates).toBeCalledTimes(1);
+            expect(applyCAPUpdates).toBeCalledWith(fs, capService, {
+                ...capProjectSettings,
+                enableNPMWorkspaces: true,
+                enableCdsUi5Plugin: true,
+                sapux: false
+            });
+        });
+
+        test('should perform CAP updates when CAP service is available, and cds ui5 plugin is disabled', async () => {
+            const fs = create(createStorage());
+
+            const capServiceWithoutCdsUi5PluginInfo = {
+                projectPath: 'test/path',
+                serviceName: 'test-service',
+                capType: 'Node.js'
+            };
+
+            const freestyleApp: FreestyleApp<BasicAppSettings> = getFreestyleApp({
+                sapux: false,
+                typescript: false,
+                capService: capServiceWithoutCdsUi5PluginInfo as CapServiceCdsInfo
+            });
+            await generate(curTestOutPath, freestyleApp, fs);
+
+            expect(applyCAPUpdates).toBeCalledTimes(1);
+            expect(applyCAPUpdates).toBeCalledWith(fs, capServiceWithoutCdsUi5PluginInfo, {
+                ...capProjectSettings,
+                sapux: false,
+                enableNPMWorkspaces: false,
+                enableCdsUi5Plugin: false
+            });
+        });
+
+        test('should not perform CAP updates, when no cap service provided', async () => {
+            const fs = create(createStorage());
+            const freestyleApp: FreestyleApp<BasicAppSettings> = getFreestyleApp({
+                sapux: false,
+                typescript: false
+            });
+            await generate(curTestOutPath, freestyleApp, fs);
+            expect(applyCAPUpdates).toBeCalledTimes(0);
         });
     });
 });