+++ examples/ui-prompting-examples/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux-private/ui-prompting-examples
 
+## 0.2.28
+
+### Patch Changes
+
+-   Updated dependencies [dd34d62]
+    -   @sap-ux/ui-components@1.24.2
+    -   @sap-ux/ui-prompting@0.3.14
+
 ## 0.2.27
 
 ### Patch Changes