+++ packages/cap-config-writer/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @sap-ux/cap-config-writer
 
+## 0.9.2
+
+### Patch Changes
+
+-   Updated dependencies [096b021]
+    -   @sap-ux/project-access@1.29.3
+    -   @sap-ux/fiori-generator-shared@0.7.23
+
+## 0.9.1
+
+### Patch Changes
+
+-   @sap-ux/fiori-generator-shared@0.7.22
+
+## 0.9.0
+
+### Minor Changes
+
+-   d6b7440: Move `CapService` and `CapRuntime` types from `@sap-ux/odata-service-inquirer` to `@sap-ux/cap-config-writer` for better relevance; add CAP updates to generate functions of Fiori Elements and Freestyle writers
+
+## 0.8.4
+
+### Patch Changes
+
+-   Updated dependencies [93ef8c1]
+    -   @sap-ux/project-access@1.29.2
+    -   @sap-ux/fiori-generator-shared@0.7.21
+
+## 0.8.3
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.29.1
+-   @sap-ux/fiori-generator-shared@0.7.20
+
 ## 0.8.2
 
 ### Patch Changes